Moved lib up from components
