+++ trekbag/src/components/Sidebar.jsx
@@ -1,6 +1,6 @@
 import AddItemForm from "./AddItemForm";
 import ButtonGroup from "./ButtonGroup";
-import { useItemsContext } from "./lib/hooks";
+import { useItemsContext } from "../lib/hooks";
 
 export default function Sidebar() {
   const { handleAddItem } = useItemsContext();
