+++ trekbag/src/lib/constants.js
