+++ trekbag/src/lib/hooks.js
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { ItemsContext } from "../../contexts/ItemsContextProvider";
+import { ItemsContext } from "../contexts/ItemsContextProvider";
 
 export function useItemsContext() {
   const context = useContext(ItemsContext);
